consultation_app/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-config-url==0.1.1
Django==5.0.6
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
django-secure==1.0.2
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
packaging==24.1
phonenumbers==8.12.48
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sentry-sdk==2.10.0
setuptools==70.3.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
